Refactor to reduce WTFs per minute in code review
When I first wrote this code back in August for the mini-hackathon it
was the first time I was going from 0-to-1 with a greenfield React
project without the crutch of existing code to reference. It is now
roughly 6 months since then and my eyes bled when I reviewed this code.
Granted this was originally built in about 4-6 hours for the LWJ
retro game mini-hackathon, so there were definitely some corner-cutting
involved in getting it working, but there was just a lot of poorly
written React in here. This is the first iteration of improvements.